requirements.txt
kafka-python==2.0.2
python-schema-registry-client==2.4.3
# kafka-schema-registry